Enhancement: Preplanner
Add a preplan capability that can search a Thunderbird profile
for common paths. This assumes the Thunderbird profile user has
multiple folders within the path that will match up and be useful
based on the mailbox structure.

By default it uses the `Inbox.sbd` folder as a split; however, the
user is able to change that via the `--pattern` parameter on the
command-line.